refactor: Create sf_ui crate for UI tools
