+++ Cargo.toml
@@ -1,18 +1,20 @@
-[package]
-name = "bevy_saltius_finni"
-authors = ["Galdormin <theo.plenet@gmail.com>"]
+[workspace]
+members = ["sf_ui"]
+
+[workspace.package]
 version = "0.1.0"
 edition = "2024"
+description = "2D MetroidVania game about an alien that dies after a few jumps."
+repository = "https://github.com/Galdormin/bevy-saltius-finni"
+authors = ["Galdormin <theo.plenet@gmail.com>"]
 
-[dependencies]
+[workspace.dependencies]
 bevy = { version = "0.16.1", features = ["wayland"] }
-rand = "0.9.1"
-# Compile low-severity logs out of native builds for performance.
+rand = { version = "0.9.1" }
 log = { version = "0.4", features = [
     "max_level_debug",
     "release_max_level_warn",
 ] }
-# Compile low-severity logs out of web builds for performance.
 tracing = { version = "0.1", features = [
     "max_level_debug",
     "release_max_level_warn",
@@ -27,45 +29,61 @@ serde = "1.0.219"
 bevy_cobweb_ui = { version = "0.20.1", default-features = false }
 bevy_cobweb = "0.17.2"
 
-# Your web builds will start failing if you add a dependency that pulls in `getrandom` v0.3+.
-# To fix this, you should tell `getrandom` to use the `wasm_js` backend on Wasm.
+
+[package]
+name = "bevy_saltius_finni"
+version.workspace = true
+edition.workspace = true
+description.workspace = true
+repository.workspace = true
+authors.workspace = true
+
+
+[dependencies]
+sf_ui = { path = "sf_ui" }
+
+bevy = { workspace = true }
+rand = { workspace = true }
+log = { workspace = true }
+tracing = { workspace = true }
+bevy_modern_pixel_camera = { workspace = true }
+avian2d = { workspace = true }
+leafwing-input-manager = { workspace = true }
+bevy_ecs_ldtk = { workspace = true }
+bevy_asset_loader = { workspace = true }
+bevy_common_assets = { workspace = true }
+serde = { workspace = true }
+bevy_cobweb_ui = { workspace = true }
+bevy_cobweb = { workspace = true }
+
+
 # See: <https://docs.rs/getrandom/0.3.3/getrandom/#webassembly-support>.
 [target.wasm32-unknown-unknown.dependencies]
 getrandom = { version = "0.3", features = ["wasm_js"] }
-# In addition to enabling the `wasm_js` feature, you need to include `--cfg 'getrandom_backend="wasm_js"'`
-# in your rustflags for both local and CI/CD web builds, taking into account that rustflags specified in
-# multiple places are NOT combined (see <https://github.com/rust-lang/cargo/issues/5376>).
-# Alternatively, you can opt out of the rustflags check with this patch:
+
 [patch.crates-io]
 getrandom = { git = "https://github.com/benfrankel/getrandom" }
 
 [features]
-# Default to a native dev build.
 default = ["dev_native"]
 dev = [
-    # Improve compile times for dev builds by linking Bevy as a dynamic library.
     "bevy/dynamic_linking",
     "bevy/bevy_dev_tools",
     "bevy/bevy_ui_debug",
-    # Improve error messages coming from Bevy
     "bevy/track_location",
 ]
 dev_native = [
     "dev",
-    # Enable asset hot reloading for native dev builds.
     "bevy/file_watcher",
-    # Enable embedded asset hot reloading for native dev builds.
     "bevy/embedded_watcher",
     "bevy_cobweb_ui/hot_reload",
 ]
 
 
 [package.metadata.bevy_cli.release]
-# Disable dev features for release builds.
 default-features = false
 
 [package.metadata.bevy_cli.web]
-# Disable native features for web builds.
 default-features = false
 
 [package.metadata.bevy_cli.web.dev]
@@ -73,19 +91,13 @@ features = ["dev"]
 
 
 [lints.rust]
-# Mark `bevy_lint` as a valid `cfg`, as it is set when the Bevy linter runs.
 unexpected_cfgs = { level = "warn", check-cfg = ["cfg(bevy_lint)"] }
 
 [lints.clippy]
-# Bevy supplies arguments to systems via dependency injection, so it's natural for systems to
-# request more than 7 arguments, which would undesirably trigger this lint.
 too_many_arguments = "allow"
-# Queries may access many components, which would undesirably trigger this lint.
 type_complexity = "allow"
-# Make sure macros use their standard braces, such as `[]` for `bevy_ecs::children!`.
 nonstandard_macro_braces = "warn"
 
-# You can configure the warning levels of Bevy lints here. For a list of all lints, see:
 # <https://thebevyflock.github.io/bevy_cli/bevy_lint/lints/>
 [package.metadata.bevy_lint]
 # panicking_methods = "deny"
