+++ src/menus/mod.rs
@@ -9,8 +9,6 @@ mod settings;
 use bevy::prelude::*;
 
 pub(super) fn plugin(app: &mut App) {
-    app.init_state::<Menu>();
-
     app.add_plugins((
         credits::plugin,
         main::plugin,
@@ -18,16 +16,4 @@ pub(super) fn plugin(app: &mut App) {
         pause::plugin,
         death::plugin,
     ));
-}
-
-#[derive(States, Copy, Clone, Eq, PartialEq, Reflect, Hash, Debug, Default)]
-#[states(scoped_entities)]
-pub enum Menu {
-    #[default]
-    None,
-    Main,
-    Credits,
-    Settings,
-    Pause,
-    Death,
 }
