+++ src/player/death.rs
@@ -1,10 +1,13 @@
 //! Handle death of player
 
+use bevy::prelude::*;
+
 use avian2d::{
     math::Vector,
     prelude::{Collider, CollisionLayers, RigidBody, SleepingDisabled},
 };
-use bevy::prelude::*;
+
+use sf_ui::prelude::Screen;
 
 use crate::{
     GameLayer,
@@ -13,7 +16,6 @@ use crate::{
         movement::JumpAmount,
         physics::{CharacterController, Grounded},
     },
-    screens::Screen,
 };
 
 pub(super) fn plugin(app: &mut App) {
