+++ src/screens/loading.rs
@@ -3,7 +3,9 @@
 
 use bevy::prelude::*;
 
-use crate::{screens::Screen, ui::prelude::*};
+use sf_ui::prelude::Screen;
+
+use crate::ui::prelude::*;
 
 pub(super) fn plugin(app: &mut App) {
     app.add_systems(OnEnter(Screen::Loading), spawn_loading_screen);
