+++ src/screens/mod.rs
@@ -8,23 +8,10 @@ mod title;
 use bevy::prelude::*;
 
 pub(super) fn plugin(app: &mut App) {
-    app.init_state::<Screen>();
-
     app.add_plugins((
         gameplay::plugin,
         loading::plugin,
         splash::plugin,
         title::plugin,
     ));
-}
-
-/// The game's main screen states.
-#[derive(Reflect, States, Copy, Clone, Eq, PartialEq, Hash, Debug, Default)]
-#[states(scoped_entities)]
-pub enum Screen {
-    #[default]
-    Splash,
-    Loading,
-    Title,
-    Gameplay,
 }
