+++ src/screens/splash.rs
@@ -2,7 +2,9 @@
 
 use bevy::{input::common_conditions::input_just_pressed, prelude::*};
 
-use crate::{AppSystems, screens::Screen, ui::prelude::*};
+use sf_ui::prelude::Screen;
+
+use crate::{AppSystems, ui::prelude::*};
 
 pub(super) fn plugin(app: &mut App) {
     // Spawn splash screen.
